Refactor: Home 컴포넌트에서 Section분리"
- styled-componrnts 이름, 태그 수정
- 디자인 일부 수정
